refactor: split nzbdav_api parsing helpers into nzbdav_api_parsing.py (iter26)
Move 20 pure response-parsing/shape/error-interpretation helpers + the _CompletedJobs
class out of nzbdav_api.py into a new sibling module nzbdav_api_parsing.py, re-exported
via __all__ so resources.lib.nzbdav_api.<name> still resolves. nzbdav_api.py NLOC
593->444 (under 500).

Patch-safe by construction: only functions that NO test @patches (and that depend
solely on never-patched names) were moved; all patched names (_get_settings, _http_get,
_history_slots, get_job_history, find_*_by_name, cancel_job, ...) stayed in nzbdav_api.py,
so every @patch("resources.lib.nzbdav_api.*") still intercepts. New module does not import
nzbdav_api (no cycle). No behavior change; no test files edited.

Codacy complexity 30 -> 29 (file-nloc>500: 7->6). Verified: 2140 tests pass,
pylint 10/10, ruff+black clean, vermin 3.8 OK.
